[update] round Admin 기능 분리
