+++ src/main/java/com/sparta/projectblue/domain/round/controller/RoundController.java
@@ -32,33 +32,5 @@ public class RoundController {
         return ResponseEntity.ok(ApiResponse.success(roundService.getAvailableSeats(id)));
     }
 
-    @PostMapping
-    @Operation(summary = "공연별 다건 회차 등록", description = "공연별 다건 회차를 등록합니다.")
-    public ResponseEntity<ApiResponse<?>> create(
-            @Valid @RequestBody CreateRoundRequestDto request) {
 
-        return ResponseEntity.ok(
-                ApiResponse.success(roundService.create(request.getPerformanceId(), request)));
-    }
-
-    @PutMapping("/{id}")
-    @Operation(summary = "회차 수정", description = "특정 회차의 정보를 수정합니다.")
-    public ResponseEntity<ApiResponse<?>> update(
-            @PathVariable Long id,
-            @RequestParam LocalDateTime date,
-            @RequestParam PerformanceStatus status) {
-
-        return ResponseEntity.ok(
-                ApiResponse.success(
-                        roundService.update(id, new UpdateRoundRequestDto(date, status))));
-    }
-
-    @DeleteMapping("/{id}")
-    @Operation(summary = "회차 삭제", description = "특정 회차를 삭제합니다.")
-    public ResponseEntity<ApiResponse<?>> delete(@PathVariable Long id) {
-
-        roundService.delete(id);
-
-        return ResponseEntity.ok(ApiResponse.successWithNoContent());
-    }
 }
